Baikal: moved functions to infrastructure; code cleanup
