+++ nrpy/infrastructures/ETLegacy/general_relativity/RegisterSlicing.py
@@ -1,0 +1,53 @@
+from typing import Union, cast, List
+from inspect import currentframe as cfr
+from types import FrameType as FT
+import sympy
+
+import nrpy.c_codegen as ccg
+import nrpy.c_function as cfc
+import nrpy.grid as gri
+import nrpy.params as par
+import nrpy.indexedexp as ixp
+import nrpy.helpers.parallel_codegen as pcg
+
+def register_CFunction_RegisterSlicing(
+    thorn_name: str,
+    ) -> Union[None, pcg.NRPyEnv_type]:
+    """
+    Register the slicing registration function.
+
+    :param thorn_name: The Einstein Toolkit thorn name.
+
+    :return: None if in registration phase, else the updated NRPy environment.
+    """
+    if pcg.pcg_registration_phase():
+        pcg.register_func_call(f"{__name__}.{cast(FT, cfr()).f_code.co_name}", locals())
+        return None
+
+    desc = f"""Register slicing condition for NRPy+-generated thorn {thorn_name}."""
+    name = f"{thorn_name}_RegisterSlicing"
+    body = f"""
+    Einstein_RegisterSlicing ("{thorn_name}");
+  return 0;"""
+
+    schedule = """
+schedule FUNC_NAME at STARTUP
+{
+  LANG: C
+  OPTIONS: meta
+} "Register 3+1 slicing condition"
+"""
+
+
+    cfc.register_CFunction(
+        subdirectory=thorn_name,
+        includes=["Slicing.h", "cctk.h"],
+        desc=desc,
+        c_type="int",
+        name=name,
+        params="",
+        body=body,
+        ET_thorn_name=thorn_name,
+        ET_schedule_bins_entries=[("STARTUP", schedule)],
+    )
+    return cast(pcg.NRPyEnv_type, pcg.NRPyEnv())
